statistics: extract the common function to rebuild the queue
